Moved all Locality related value objects to separate dir
