+++ src/Value/Locality/LocalityNameId.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigipolisGent\Flanders\BasicRegisters\Value\Locality;
+
+use DigipolisGent\Flanders\BasicRegisters\Value\AbstractId;
+
+/**
+ * Locality name ID value.
+ */
+final class LocalityNameId extends AbstractId
+{
+
+}
