+++ src/Value/Locality/LocalityNames.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Flanders\BasicRegisters\Value;
+namespace DigipolisGent\Flanders\BasicRegisters\Value\Locality;
 
 use DigipolisGent\Value\CollectionAbstract;
 
@@ -14,7 +14,7 @@ final class LocalityNames extends CollectionAbstract
     /**
      * Create a new collection.
      *
-     * @param \DigipolisGent\Flanders\BasicRegisters\Value\LocalityName ...$localityNames
+     * @param \DigipolisGent\Flanders\BasicRegisters\Value\Locality\LocalityName ...$localityNames
      *   One or more locality names.
      */
     public function __construct(LocalityName ...$localityNames)
@@ -29,7 +29,7 @@ final class LocalityNames extends CollectionAbstract
     {
         $localityNames = [];
 
-        /** @var \DigipolisGent\Flanders\BasicRegisters\Value\LocalityName $localityName */
+        /** @var \DigipolisGent\Flanders\BasicRegisters\Value\Locality\LocalityName $localityName */
         foreach ($this->values as $localityName) {
             $localityNames[] = (string) $localityName;
         }
