--- src/Value/LocalityNameId.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DigipolisGent\Flanders\BasicRegisters\Value;
-
-/**
- * Locality name ID value.
- */
-final class LocalityNameId extends AbstractId
-{
-
-}
