+++ tests/Normalize/FromJson/StreetNameDetailNormalizerTest.php
@@ -8,8 +8,8 @@ use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\StreetNameDetailNo
 use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName;
 use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
-use DigipolisGent\Flanders\BasicRegisters\Value\LocalityName;
-use DigipolisGent\Flanders\BasicRegisters\Value\LocalityNameId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Locality\LocalityName;
+use DigipolisGent\Flanders\BasicRegisters\Value\Locality\LocalityNameId;
 use DigipolisGent\Flanders\BasicRegisters\Value\StreetNameDetail;
 use DigipolisGent\Flanders\BasicRegisters\Value\StreetNameId;
 use PHPUnit\Framework\TestCase;
